+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_document.rb
@@ -4,10 +4,18 @@ class RubyLsp::Ree::ParsedDocument
   LINK_DSL_MODULE = 'Ree::LinkDSL'
 
   attr_reader :ast, :package_name, :class_node, :fn_node, :fn_block_node, :class_includes,
-    :link_nodes, :values
+    :link_nodes, :values, :action_node, :action_block_node
 
   def initialize(ast)
     @ast = ast
+  end
+
+  def links_container_node
+    @fn_node || @action_node
+  end
+
+  def links_container_block_node
+    @fn_block_node || @action_block_node
   end
 
   def includes_link_dsl?
@@ -22,8 +30,8 @@ class RubyLsp::Ree::ParsedDocument
     @link_nodes.map(&:name).include?(obj_name)
   end
 
-  def has_blank_fn?
-    @fn_node && !@fn_block_node
+  def has_blank_links_container?
+    links_container_node && !links_container_block_node
   end
 
   def set_package_name(package_name)
@@ -39,6 +47,13 @@ class RubyLsp::Ree::ParsedDocument
 
     @fn_node ||= class_node.body.body.detect{ |node| node.name == :fn }
     @fn_block_node = @fn_node&.block
+  end
+
+  def parse_action_node
+    return unless class_node
+
+    @action_node ||= class_node.body.body.detect{ |node| node.name == :action }
+    @action_block_node = @action_node&.block
   end
 
   def parse_class_includes
@@ -57,8 +72,8 @@ class RubyLsp::Ree::ParsedDocument
   def parse_links
     return unless class_node
 
-    nodes = if fn_block_node && fn_block_node.body
-      fn_block_node.body.body.select{ |node| node.name == :link }
+    nodes = if links_container_node && links_container_block_node.body
+      links_container_block_node.body.body.select{ |node| node.name == :link }
     elsif class_includes.any?{ _1.name == LINK_DSL_MODULE }
       class_node.body.body.select{ |node| node.name == :link }
     else