+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/parsing/parsed_document_builder.rb
@@ -32,6 +32,7 @@ class RubyLsp::Ree::ParsedDocumentBuilder
     
     document.parse_class_node
     document.parse_fn_node
+    document.parse_action_node
     document.parse_class_includes
     document.parse_links